more fixy

+++ src/cli_args.rs
@@ -8,6 +8,6 @@ use clap::Parser;
 #[clap(author = "runtime", version, about, long_about = None)]
 pub struct CliArgs {
     /// Sets the level of verbosity
-    #[clap(short, long, parse(from_occurrences))]
-    pub verbose: usize,
+    #[clap(short, long, action = clap::ArgAction::Count)]
+    pub verbose: u8,
 }
